feat(knowledge): unify Manuals, Map and Suggestions under one Knowledge tab
Fold the standalone /graph and /proposals routes into the Knowledge section
as sub-tabs — Manuals (KB) · Map (relationship graph) · Suggestions
(propose/verify queue) — collapsing three sidebar items into one.

- (hub)/knowledge/layout.tsx + KnowledgeTabs.tsx: sub-tab bar with pending
  suggestion count; default sub-tab is Map; plain-English labels for techs
- reparent pages (git mv): knowledge->manuals, graph->map, proposals->suggestions
- legacy routes redirect in: /graph and /proposals (graph preserves the
  reasoning-trace deep link ?session=...), /library and /knowledge -> manuals/map
- nav consolidated to a single "Knowledge" primary item
- 6 vitest cases: redirects (incl. ?session= preservation) + nav consolidation

No backend or flow change; never-auto-verify preserved.
