+++ mira-hub/src/app/(hub)/library/page.tsx
@@ -1,8 +1,8 @@
 import { redirect } from "next/navigation";
 
 // /library was the standalone KB browser. Library now lives inside the
-// Knowledge tab — keep the route as a permanent redirect so any saved
-// link, bookmark, or QR code still resolves.
+// Knowledge section as the "Manuals" sub-tab — keep the route as a permanent
+// redirect so any saved link, bookmark, or QR code still resolves.
 export default function LegacyLibraryRedirect() {
-  redirect("/knowledge");
+  redirect("/knowledge/manuals");
 }
